UML-32 refactor ComponentDiagram struct
