+++ src/components/ComponentDiagram/Diagram/__tests__/Diagram.test.ts
@@ -5,7 +5,7 @@ import messagesEn from '../../../../locales/en/messages';
 import Diagram from '../index';
 import ComponentType from '../../../../models/ComponentType';
 import ComponentFactory from '../../../../models/factories/ComponentFactory';
-import LinkValidator from '../models/LinkValidator';
+import LinkValidator from '../../elements/LinkValidator';
 
 const componentFactory = new ComponentFactory();
 const linkValidator = new LinkValidator();
