+++ src/components/ComponentDiagram/actions/ZoomAction.ts
@@ -1,6 +1,6 @@
 import { Action, InputType } from '@projectstorm/react-canvas-core';
 import { WheelEvent } from 'react';
-import isType from '../../../../utils/guards/isType';
+import isType from '../../../utils/guards/isType';
 
 const ZOOM_LEVELS = [10, 25, 50, 75, 100, 150, 200, 300];
