+++ src/components/ComponentDiagram/elements/DiagramCanvas/DiagramCanvas.module.scss
