+++ src/components/ComponentDiagram/elements/Link/Link.ts
