+++ src/components/ComponentDiagram/elements/Link/NodeLinkWidget.tsx
