+++ src/components/ComponentDiagram/elements/Node/NodeClass/NodeClassFactory.tsx
@@ -2,14 +2,14 @@ import * as React from 'react';
 import { AbstractReactFactory } from '@projectstorm/react-canvas-core';
 import { DiagramEngine } from '@projectstorm/react-diagrams-core';
 import { PortModel } from '@projectstorm/react-diagrams';
-import { Node } from '../models/Node';
-import NodeInterfaceWidget from '../widgets/NodeWidgets/NodeInterfaceWidget';
-import ComponentType from '../../../../models/ComponentType';
-import { COMPONENTS_NAMES } from '../../../../locales/lang-constants';
-import ComponentFactory from '../../../../models/factories/ComponentFactory';
-import { LinkValidatorI } from '../models/LinkValidator';
+import { Node } from '../Node';
+import NodeClassWidget from './NodeClassWidget';
+import ComponentType from '../../../../../models/ComponentType';
+import { COMPONENTS_NAMES } from '../../../../../locales/lang-constants';
+import ComponentFactory from '../../../../../models/factories/ComponentFactory';
+import { LinkValidatorI } from '../../LinkValidator';
 
-type NodeInterfaceFactoryProps = {
+type NodeClassFactoryProps = {
     factory: ComponentFactory
     linkValidator: LinkValidatorI
     findConnection(): {
@@ -18,22 +18,22 @@ type NodeInterfaceFactoryProps = {
 };
 
 /**
- * Factory to create the "Interface" nodes of the diagram.
+ * Factory to create the "Class" nodes of the diagram.
  * With its help the widget of such a node is created.
  */
-export default class NodeInterfaceFactory extends AbstractReactFactory<Node, DiagramEngine> {
-    private readonly componentType = ComponentType.INTERFACE;
+export class NodeClassFactory extends AbstractReactFactory<Node, DiagramEngine> {
+    private readonly componentType = ComponentType.CLASS;
 
     private readonly factory: ComponentFactory;
 
     private readonly linkValidator: LinkValidatorI;
 
-    private findConnection: () => {
+    private readonly findConnection: () => {
         port: null | PortModel,
     };
 
-    constructor(props: NodeInterfaceFactoryProps) {
-        super(ComponentType.INTERFACE);
+    constructor(props: NodeClassFactoryProps) {
+        super(ComponentType.CLASS);
         this.factory = props.factory;
         this.linkValidator = props.linkValidator;
         this.findConnection = props.findConnection;
@@ -41,10 +41,10 @@ export default class NodeInterfaceFactory extends AbstractReactFactory<Node, Dia
 
     generateReactWidget(event: { model: Node }): JSX.Element {
         return (
-            <NodeInterfaceWidget
+            <NodeClassWidget
+                findConnection={this.findConnection}
                 engine={this.engine}
                 node={event.model}
-                findConnection={this.findConnection}
             />
         );
     }
