+++ src/components/ComponentDiagram/elements/Node/NodeWidget/Content/Content.module.scss
