+++ src/components/ComponentDiagram/elements/Node/NodeWidget/NodeWidget.module.scss
