+++ src/components/ComponentDiagram/elements/Port/Port.ts
@@ -5,10 +5,10 @@ import {
 } from '@projectstorm/react-diagrams';
 import { BaseEntityEvent, BaseEvent } from '@projectstorm/react-canvas-core';
 import isType from '../../../../utils/guards/isType';
-import { LinkValidatorI } from './LinkValidator';
+import { LinkValidatorI } from '../LinkValidator';
 import Observable from '../../../../lib/Observable';
-import { NodeI } from './Node';
-import Link from './Link';
+import { NodeI } from '../Node/Node';
+import Link from '../Link/Link';
 
 export type PortEvent = BaseEntityEvent<LinkModel> & { port: null | PortModel };
